GUILOAD: don't abandon the current game until we are sure
unlike the originals, we don't have to throw out the current game until we actually need to load a new one
this allows you to see the list of saves without stopping the game and you can cancel to return to the existing game
